weightb: add CLI options for explicit weightb [CHANGES OUTPUTS]
weightB will be disabled by default for this release. This patch fixes the hash
mistakes and other fetaure bugs (the reasons it had been disabled)

HEVC only signals a single denom for all references so the L1 reference must use
the same denom as the L0 reference or no weight at all. A more exhaustive
search might be added later for slower presets. Also, the lookahead will need to
be modified to model weightB behavior so it can be more effective.

This patch changes how the unweighted references are coded, so even weightP
outputs change slightly.