param: fix profile warnings in x265_param_apply_profile()
This function was likely never used since it was not supported by the CLI till
now.  When the user specifies --profile Main10, they intend that the output
stream is decodable by a Main10 decoder. This is true of any Main encode.

The reverse is not true for --profile Main and a Main10 encode.  We must return
an error and abort the encode if --profile Main is specified and the encoder
was compiled for Main10 (HIGH_BIT_DEPTH).