bitcost: fix bitcost() function, broken 13 months ago by 07015bbe306b
when bitcost() was first introduced, we had two competing table formats for the
s_bitsizes[] array. Eventually the x264 style array won out, but this function
was not updated.

This bug was almost harmless, bitcost() could still evaluate differences in bit
costs correctly, but the upshift caused read bound exceptions when the motion
vectors became close to the HEVC max of 4k.