analysis: move encodeCU() methods and friends to Entropy
This function only encodes a fully finished CTU; it is performing no analysis
and it needs none of the features or members of the Analysis class. It is much
more clear as a member of the Entropy class.

Note: the way bEncodeDQP is communicated is a huge mess and needs to be cleaned
up soon. And finishCU() can also likely be hugely simplified