api: improve error handling in x265_api_get() bit depth checks
These checks were added to ensure that the returned API supported the bitdepth
that is implied by the library's filename (libx265_main10 should support main10)
but the checks themselves were based on a broken premise. If the file was
misnamed, then it will return NULL if you ask for its implied target bitDepth.

To fix this we always call x265_api_get(0) for the bound library, which must
always return a non-NULL pointer, and then verify that the returned API's bit
depth indeed matches the one implied by its filename.