param: Disable opt-qp-pps and opt-ref-list-length-pps by default
The bits saved by these options is very miniscule, and not observable in most
tests focusing on low bitrates, and low resolutions; this was the goal of these
two options. Also, since these options change the PPS, the muxer has to ensure
that the headers are sent every frame which seems to trouble some muxers and
users.