Clean up SEI::write function
define another SEI function to count the payload size and return count in bits.
Set the payload size before calling SEI::write() to avoid clutter of if conditions in
write function. Write payloadtype and payload size in bitstream for all SEI in same way
as it is the syntax defined in the spec for any SEI.